Rename module app_ to callbacks
