+++ dve/callbacks/colour_scale.py
